CJK Compatibilty
Contains a lot of 2 and 3 letter combinations, scientific measuring
units, each with their own casing. So MHZ becomes MHz, PV becomes pV.
The problem is that some of these patterns overlap and it is a bit of a
mess getting it all to line up.  So for the time being, all units as
compressed, lowercase names.